+++ src/BitcoinKernel.Interop/Delegates/Notification/NotifyFlushError.cs
@@ -4,6 +4,6 @@ using System.Runtime.InteropServices;
 namespace BitcoinKernel.Interop.Delegates.Notification;
 
 [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-public delegate void NotifyFlushError(
+internal delegate void NotifyFlushError(
     IntPtr user_data,
     [MarshalAs(UnmanagedType.LPUTF8Str)] string error_message);