+++ src/BitcoinKernel.Interop/Delegates/Notification/NotifyWarningSet.cs
@@ -4,6 +4,6 @@ using BitcoinKernel.Interop.Enums;
 namespace BitcoinKernel.Interop.Delegates.Notification;
 
 [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-public delegate void NotifyWarningSet(
+internal delegate void NotifyWarningSet(
     IntPtr user_data,
     Warning warning);