+++ src/BitcoinKernel.Interop/Delegates/Validation/ValidationBlockDisconnected.cs
@@ -3,6 +3,6 @@ using System.Runtime.InteropServices;
 namespace BitcoinKernel.Interop.Delegates.Validation;
 
 [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-public delegate void ValidationBlockDisconnected(
+internal delegate void ValidationBlockDisconnected(
     IntPtr user_data,
     IntPtr block);