+++ src/BitcoinKernel.Interop/Delegates/Validation/ValidationNewPoWValidBlock.cs
@@ -3,7 +3,7 @@ using System.Runtime.InteropServices;
 namespace BitcoinKernel.Interop.Delegates.Validation;
 
 [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-public delegate void ValidationNewPoWValidBlock(
+internal delegate void ValidationNewPoWValidBlock(
     IntPtr user_data,
     IntPtr block_index,
     IntPtr block);