+++ src/BitcoinKernel.Interop/Enums/ChainType.cs
@@ -1,11 +1,10 @@
-namespace BitcoinKernel.Interop.Enums
+namespace BitcoinKernel.Interop.Enums;
+
+public enum ChainType : uint
 {
-    public enum ChainType : uint
-    {
-        MAINNET = 0,
-        TESTNET = 1,
-        TESTNET_4 = 2,
-        SIGNET = 3,
-        REGTEST = 4
-    }
+    MAINNET = 0,
+    TESTNET = 1,
+    TESTNET_4 = 2,
+    SIGNET = 3,
+    REGTEST = 4
 }