+++ src/BitcoinKernel.Interop/Structs/LoggingOptions.cs
@@ -1,14 +1,13 @@
 using System.Runtime.InteropServices;
 
-namespace BitcoinKernel.Interop.Structs
+namespace BitcoinKernel.Interop.Structs;
+
+[StructLayout(LayoutKind.Sequential)]
+internal struct LoggingOptions
 {
-    [StructLayout(LayoutKind.Sequential)]
-    public struct LoggingOptions
-    {
-        public int LogTimestamps;
-        public int LogTimeMicros;
-        public int LogThreadNames;
-        public int LogSourceLocations;
-        public int AlwaysPrintCategoryLevels;
-    }
+    public int LogTimestamps;
+    public int LogTimeMicros;
+    public int LogThreadNames;
+    public int LogSourceLocations;
+    public int AlwaysPrintCategoryLevels;
 }